* AssetManager now an interface  * Must use G3DSystem.newAssetManager() to create asset manager.  * BinaryImporter/BinaryExporter now reusable if used in same thread  * Better fixed-pipeline support: Material will load FixedFunc technique, ortho setup in RenderManager for Gui queue.  * Renderer.setView and setProj methods merged  * VertexBuffer supports byte data  * Image got android formats and paletted formats

+++ branches/jme3/src/jme3_core/com/g3d/scene/dbg/Arrow.java
@@ -41,7 +41,7 @@ public class Arrow extends Mesh {
 
         setBuffer(Type.Position, 3, newPositions);
         setBuffer(Type.Index, 2,
-                new int[]{
+                new short[]{
                     0, 1,
                     1, 2,
                     1, 3,
